Configure database location in config
