+++ config.toml
@@ -5,6 +5,7 @@ user_config_category = "notify"
 wiki_config_category = "wiki"
 overrides_url = "https://notifications.wdfiles.com/local--code/overrides/1"
 gmail_username = "wikidotnotifier@gmail.com"
+database_driver = "notifier.database.drivers.sqlite.SqliteDriver"
 
 [notifier.path]
 # Paths:
