--- notifier/database/drivers/__init__.py
@@ -1,3 +1,0 @@
-from notifier.database.drivers.sqlite import SqliteDriver
-
-DatabaseDriver = SqliteDriver
